+++ src/routing/normalization/normalize-route-props.ts
@@ -1,7 +1,7 @@
 import type { HttpRequest } from '@/Http';
+import type { HttpMethod } from '@/routing/declaration/http-method.type';
 import type { NormalizedRouteProps } from '@/routing/declaration/normalized-route-props.type';
 import { RouteProps } from '@/routing/declaration/route-props.type';
-import type { HttpMethod } from '@/routing/registration/verb/http-method.type';
 import type { RouteOptions } from '@/routing/route/route-options.type';
 import type { RouterMethod } from '@/routing/route/router-method.type';
 import type { Request } from 'koa';
